build: allow build scripts to bail out in case the lib is already available

+++ scripts/ci/build-zlib.sh
@@ -8,6 +8,14 @@ curr_dirname=$(dirname "$0")
 mkdir -p $build_folder
 mkdir -p $2/source
 
+FORCE_REBUILD=${FORCE_REBUILD:-}
+
+# @TODO We are explicitly checking the static lib
+if [[ -f $build_folder/lib/libz.a ]] && [[ -z $FORCE_REBUILD || $FORCE_REBUILD != "true" ]]; then
+  echo "Skipping rebuild of zlib because lib file already exists"
+  exit 0
+fi
+
 $curr_dirname/download-and-unpack.sh https://github.com/madler/zlib/archive/v$1.tar.gz $2
 
 mv $2/zlib-$1 $2/source/$1
